+++ operate/client/src/modules/bpmn-js/badgePositions.ts
@@ -39,6 +39,11 @@ const SUBPROCESS_WITH_INCIDENTS = {
   right: -12,
 };
 
+const WAITING_BADGE = {
+  top: -36,
+  left: 6,
+};
+
 export {
   MODIFICATIONS,
   DECISION_STATE,
@@ -48,4 +53,5 @@ export {
   COMPLETED_BADGE,
   COMPLETED_END_EVENT_BADGE,
   SUBPROCESS_WITH_INCIDENTS,
+  WAITING_BADGE,
 };